+++ GameOfLife/src/GoL_Engine.cpp
@@ -7,26 +7,26 @@
 #include <omp.h>
 
 namespace{
-	inline size_t countAliveNeighbours(size_t x, size_t y, const Grid& grid){
+	inline size_t countAliveNeighbours(int x, int y, const Grid& grid){
 		return grid(x - 1, y - 1).isAlive() + grid(x - 1, y).isAlive() + grid(x - 1, y + 1).isAlive() +
 			grid(x, y - 1).isAlive() + grid(x, y + 1).isAlive() +
 			grid(x + 1, y - 1).isAlive() + grid(x + 1, y).isAlive() + grid(x + 1, y + 1).isAlive();
 	}
 
-	inline size_t countDeadNeighbours(size_t x, size_t y, const Grid& grid){
+	inline size_t countDeadNeighbours(int x, int y, const Grid& grid){
 		return 8 - countAliveNeighbours(x, y, grid);
 	}
 }
 
-GoL_Engine::GoL_Engine(size_t width_pow2, size_t height_pow2, size_t fill_thresold) :
-grid_(std::pow(2, width_pow2), std::pow(2, height_pow2)),
+GoL_Engine::GoL_Engine(int width_pow2, int height_pow2, int fill_thresold) :
+grid_(static_cast<int>(std::pow(2, width_pow2)), static_cast<int>(std::pow(2, height_pow2))),
 fill_thresold_(fill_thresold),
 step_number_(0),
 max_step_(std::numeric_limits<size_t>::max())
 {}
 
-GoL_Engine::GoL_Engine(size_t width_pow2, size_t height_pow2, size_t fill_thresold, size_t max_step) :
-grid_(std::pow(2, width_pow2), std::pow(2, height_pow2)),
+GoL_Engine::GoL_Engine(int width_pow2, int height_pow2, int fill_thresold, size_t max_step) :
+grid_(static_cast<int>(std::pow(2, width_pow2)), static_cast<int>(std::pow(2, height_pow2))),
 fill_thresold_(fill_thresold),
 step_number_(0),
 max_step_(max_step)
@@ -41,8 +41,8 @@ void GoL_Engine::init(){
 	std::mt19937 rd_mt_engine(seed); // mt19937 est le mersenne_twister_engine standard
 	std::uniform_int_distribution<int> uniform_distrib(1, 100); // distribution 1 à 100 uniforme
 
-	for (size_t i = 0; i < grid_.width(); ++i){
-		for (size_t j = 0; j < grid_.height(); ++j){
+	for (int i = 0; i < grid_.width(); ++i){
+		for (int j = 0; j < grid_.height(); ++j){
 			//Remplissage aléatoire de la grille en fonction du fill_thresold
 			if (uniform_distrib(rd_mt_engine) < fill_thresold_){
 				grid_(i, j).resurect();
@@ -58,8 +58,8 @@ const Grid* GoL_Engine::do_step(){
 	if (step_number_ < max_step_){
 		auto grid_working_cpy = Grid{ grid_.width(), grid_.height() }; //Copie de travail
 
-		for (size_t i = 0; i < grid_.width(); ++i){
-			for (size_t j = 0; j < grid_.height(); ++j){
+		for (int i = 0; i < grid_.width(); ++i){
+			for (int j = 0; j < grid_.height(); ++j){
 				const auto cells_alive = ::countAliveNeighbours(i, j, grid_);
 				//Any live cell with fewer than two live neighbors dies, as if caused by under - population.
 				//Any live cell with more than three live neighbors dies, as if by overcrowding.
@@ -128,14 +128,14 @@ const Grid* GoL_Engine::do_step_omp(){
 	}
 }
 
-const Grid* GoL_Engine::do_step_std_thread(size_t nb_thread){
+const Grid* GoL_Engine::do_step_std_thread(int nb_thread){
 	if (step_number_ < max_step_){
 		auto grid_working_cpy = Grid{ grid_.width(), grid_.height() }; //Copie de travail
 
-		const auto thread_functor_factory = [&](size_t start_width, size_t end_width, size_t start_height, size_t end_height) -> std::function<void(void)>{
+		const auto thread_functor_factory = [&](int start_width, int end_width, int start_height, int end_height) -> std::function<void(void)>{
 			return [&](){
-				for (size_t i = start_width; i < end_width && i < grid_.width(); ++i){
-					for (size_t j = start_height; j < end_height && j < grid_.height(); ++j){
+				for (int i = start_width; i < end_width && i < grid_.width(); ++i){
+					for (int j = start_height; j < end_height && j < grid_.height(); ++j){
 						const auto cells_alive = ::countAliveNeighbours(i, j, grid_);
 						//Any live cell with fewer than two live neighbors dies, as if caused by under - population.
 						//Any live cell with more than three live neighbors dies, as if by overcrowding.
@@ -155,12 +155,12 @@ const Grid* GoL_Engine::do_step_std_thread(size_t nb_thread){
 			};
 		};
 
-		const size_t delta_width = grid_.width() / (std::sqrt(nb_thread));
-		const size_t delta_height = grid_.height() / (std::sqrt(nb_thread));
+		const int delta_width = static_cast<int>(grid_.width() / std::sqrt(nb_thread));
+		const int delta_height = static_cast<int>(grid_.height() / std::sqrt(nb_thread));
 		std::vector<std::thread> thread_list;
 
-		for (size_t i = 0; i < grid_.width(); i += delta_width){
-			for (size_t j = 0; j < grid_.height(); j += delta_height){
+		for (int i = 0; i < grid_.width(); i += delta_width){
+			for (int j = 0; j < grid_.height(); j += delta_height){
 				thread_list.emplace_back(std::move(std::thread{ thread_functor_factory(i, i + delta_width, j, j + delta_height) }));
 			}
 		}
@@ -185,19 +185,19 @@ void GoL_Engine::run_serial(){
 	while (do_step() != nullptr);
 }
 
-void GoL_Engine::run_omp(size_t nb_thread){
+void GoL_Engine::run_omp(int nb_thread){
 	omp_set_num_threads(nb_thread);
 	while (do_step_omp() != nullptr);
 }
 
-void GoL_Engine::run_std_thread(size_t nb_thread){
+void GoL_Engine::run_std_thread(int nb_thread){
 	while (do_step_std_thread(nb_thread) != nullptr);
 }
 
 
 bool GoL_Engine::allAlive() const{
-	for (size_t i = 0; i < grid_.width(); ++i){
-		for (size_t j = 0; j < grid_.height(); ++j){
+	for (int i = 0; i < grid_.width(); ++i){
+		for (int j = 0; j < grid_.height(); ++j){
 			if (grid_(i, j).isDead())
 				return false;
 		}
@@ -206,8 +206,8 @@ bool GoL_Engine::allAlive() const{
 }
 
 bool GoL_Engine::allDead() const{
-	for (size_t i = 0; i < grid_.width(); ++i){
-		for (size_t j = 0; j < grid_.height(); ++j){
+	for (int i = 0; i < grid_.width(); ++i){
+		for (int j = 0; j < grid_.height(); ++j){
 			if (grid_(i, j).isAlive())
 				return false;
 		}
@@ -229,8 +229,8 @@ std::ostream& operator<<(std::ostream& oss, const GoL_Engine& gol_engine) {
 	auto grid = gol_engine.getGrid();
 	oss << "Grille " << grid.width() << "x" << grid.height() << std::endl;
 	oss << "----" << std::endl;
-	for (size_t i = 0; i < grid.width(); ++i){
-		for (size_t j = 0; j < grid.height(); ++j){
+	for (int i = 0; i < grid.width(); ++i){
+		for (int j = 0; j < grid.height(); ++j){
 			oss << (grid(i, j).isAlive() ? "O" : "X");
 		}
 		oss << std::endl;
